requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==2.1.0
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
fpdf==1.7.2
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
py3-validate-email==1.0.2
python-http-client==3.3.2
requests==2.25.1
sendgrid==6.7.0
SQLAlchemy==1.3.17
starkbank-ecdsa==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
